+++ docs/changelog.md
@@ -70,6 +70,8 @@ a full list of changes to the project, please consult the commit log.
       marked target, and enemies within `192` range of it, will take `60`
       spell damage per second for the next `6.0` seconds. Additional attacks
       only refreshes this duration.
+- Stone of Bryvx:
+    - Damage bonus reduced from `50%` to `30%`.
 
 ### Fixed
 - Air pathing is now more consistent across all players.